move and minor refactor BaseAdapter
